+++ src/Module/Registry.php
@@ -52,6 +52,7 @@ final class Registry {
 		\GTM4WP\Modules\Blacklist\BlacklistModule::class,
 		\GTM4WP\Modules\GoogleAuth\GoogleAuthModule::class,
 		\GTM4WP\Modules\GoogleDataManager\GoogleDataManagerModule::class,
+		\GTM4WP\Modules\Services\ServicesModule::class,
 	);
 
 	/**